SJOSearchableFetchedResultsController can be constructed with a managed object context.
This allows a SJOSearchableFetchedResultsController to be constructed without
the dependency on SJOStore. This is helpful if you only want to use this
class without the the rest of SJODataKit, say with your own Core Data
framework, or simply using standard Core Data classes.

If the `store` property is set then the `managedObjectContext` property
is set to be the `store.mainContext` managed object context.

I think that another constructed should also be offered now, and the `store`
property made readonly publicly. This way the `managedObjectContext` property
would be immutable which would better as you would't want this changing
after the object's construction.