Method can be overridden to specify a sectionKeyPath if a section index is to be displayed for the tableView.
If `searchIsActive` is YES then the sectionKeyPath return value will be
ignored and nil used regardless. This is because A a section index should
not be shown while searching, and B executed fetch requests take longer
when sections are used. When searching this is especially noticable as a
new fetch request is executed upon each key stroke during search.